Split out PlayerSummaryButton
