+++ src/main/kotlin/dartzee/screen/player/PlayerAchievementsButton.kt
@@ -8,32 +8,27 @@ import dartzee.db.PlayerEntity
 import dartzee.game.GameType
 import dartzee.screen.ScreenCache
 import dartzee.screen.stats.player.PlayerAchievementsScreen
-import java.awt.Dimension
-import java.awt.event.ActionEvent
 import java.awt.event.ActionListener
-import java.awt.event.MouseEvent
-import java.awt.event.MouseListener
 import javax.swing.ImageIcon
-import javax.swing.JButton
 
 class PlayerAchievementsButton(private val player: PlayerEntity,
-                               private val achievementRows: List<AchievementEntity>): JButton(), ActionListener, MouseListener
+                               private val achievementRows: List<AchievementEntity>): PlayerSummaryButton(), ActionListener
 {
+    override val defaultText = makeDefaultText()
+    override val hoverText = "<html><h3>Achievements &gt;</h3></html>"
+
     private val iconUrl = getGamesWonIcon(GameType.X01)
-    private val statsText = makeStatsText()
 
     init
     {
-        preferredSize = Dimension(275, 100)
-
         icon = ImageIcon(iconUrl)
-        text = statsText
+        text = defaultText
 
         addActionListener(this)
         addMouseListener(this)
     }
 
-    private fun makeStatsText(): String
+    private fun makeDefaultText(): String
     {
         val score = getPlayerAchievementScore(achievementRows, player)
         val lineOne = "<h3>Achievements</h3>"
@@ -41,31 +36,11 @@ class PlayerAchievementsButton(private val player: PlayerEntity,
         return "<html><center>$lineOne<br>$lineTwo</center></html>"
     }
 
-    override fun actionPerformed(arg0: ActionEvent)
+    override fun buttonPressed()
     {
         val scrn = ScreenCache.get<PlayerAchievementsScreen>()
         scrn.player = player
         scrn.previousScrn = ScreenCache.get<PlayerManagementScreen>()
-
         ScreenCache.switch(scrn)
-
-        text = statsText
-    }
-
-    override fun mouseClicked(e: MouseEvent?) {}
-    override fun mouseReleased(e: MouseEvent?) {}
-    override fun mousePressed(e: MouseEvent?) {}
-
-    override fun mouseEntered(e: MouseEvent?)
-    {
-        if (isEnabled)
-        {
-            text = "<html><h3>Achievements &gt;</h3></html>"
-        }
-    }
-
-    override fun mouseExited(e: MouseEvent?)
-    {
-        text = statsText
     }
 }
